fixing #41 to be able to test with files
This reduces the generated code to the interfaces. Thus, the bug with optional parameters next to files is avoided. This means that our controllers are now the actual controllers. The additionalMetadata does not show up in my Swagger UI, but it does in IntelliJ and the JS Client seems to be able to handle this as well.