Storing which model has been used to embed source
For each source entry it is now also stored, which models it has been embedded with. This way a source can be embedded with multiple models at once. For each model, the vector store will use a different table. This way old data can be dropped easily and sources can be split and embedded based on the original source if the model changes.