Avoid request timeouts when indexing
The indexing is now done in a separate thread and waited for with a timeout. This timeout is not equal to the actual request timeout and should be below that to account for the overhead of the request handling. Processing of requests is wrapped in a CompletableFuture and given to a VirtualThreadPerTaskExecutor for execution. The ExecutorService used could be moved to a more general location, if it should be used by other components as well.

The timeout can be changed via properties.

If the future times out, 202 Accepted is returned. Otherwise, a normal 200 is returned. This has the effect, that the Documents are not immediately available when searching.

It is also assumed, that when a timeout occurs, the request only has to finish embedding the documents. All errors when reading the sources should be thrown before.

Alternative:

Spring provides the classes DeferredResult and WebAsyncTask. Those classes provide functionality to return different responses, if the task times out or throws exceptions. Apparently, the executing thread is interrupted upon timout. Thus, the creation of embeddings is interrupted causing the task to fail. Thus, if we want to use those classes, we have to wrap the execution in our own thread like in the provided solution to avoid the interrupt.

There is also another problem. When using the mentioned classes, the ResponseEntity is wrapped in them. Thus, the method signature changes. The OpenAPI Generator supports that. However, there is only one flag for it, and it changes the signature for all endpoints.