WIP Make sure embeddings don't fail with rate limits
- When documents are ready to be embedded, they are now added to an EmbeddingBacklog
- The EmbeddingTaskExecutor takes document from the EmbeddingBacklog and tries to embed them.
- If this fails, it tries again after some time.

I also tested other methods to prevent hitting the rate limit. When we are using an embedding model with a larger context window, we may increase the batch size for requests and not the length of the actual documents. Thus, the chunks of test are not too large, and we still get a performance boost due to fewer calls to the model.

TODO:

- [ ] differentiate between exceptions in the EmbeddingTaskExecutor
- [ ] decide between batch size and total token count configuration properties
- [ ] maybe implement a database version of the EmbeddingBacklog
- [ ] make it possible to wait for task completion