Remove EmbeddingTasks from EmbeddingBacklog if the source is deleted
When a source is deleted, the EmbeddingBacklog is tasked to remove the associated task, if it exists.

There is still a race condition. If the task was removed while the EmbeddingTaskExecutor already has the task and is not done yet, the source could show up in the VectorStore. I am not sure yet, what to do about that.